game redreshing components independently now
