+++ frontend/src/components/Game/ComputerPaddle.module.css
